+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: lebarbos <lebarbos@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/12 15:39:30 by lebarbos          #+#    #+#             */
-/*   Updated: 2023/05/01 16:05:19 by lebarbos         ###   ########.fr       */
+/*   Updated: 2023/10/05 19:17:52 by lebarbos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@ RETURN VALUE: The converted value.
 
 #include "libft.h" 
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	result;
-	int	sign;
+	long	result;
+	int		sign;
 
 	result = 0;
 	sign = 1;
